% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_output.R
\docType{data}
\name{example_output}
\alias{example_output}
\title{example_output}
\format{
A dataframe containing the columns described below.
}
\source{
A complete small-scale simulation study was run using the functions
in \code{ValidationExplorer}.
}
\description{
A dataframe in the format output from running \link{run_sims}. This exists
solely for the purpose of making documentation of visualization functions
easier.
\itemize{
\item{parameter. Parameters (lambda, psi, and theta for a two species assemblage).}
\item{Mean. The posterior mean.}
\item{SD. The standard deviation of the draws of reach parameter.}
\item{Naive SE. The standard error, not accounting for the correlation in draws}
\item{Time-series SE. The standard error, accounting for correlation in draws.}
\item{quantiles. 2.5\%, 25\%, 50\%, 75\% and 97.5\% quantiles.}
\item{Rhat. The Gelman-Rubin statistic for each parameter.}
\item{ess_bulk. The effective sample size in the bulk of the distribution.}
\item{ess_tail. The effective sample size in the tails of the distribution.}
\item{truth. The known true data generating value.}
\item{capture. Did the 95\% posterior interval contain the true value?}
\item{converge. Was the Gelman-Rubin statistic near 1?}
\item{theta_scenario. The ID for the classifier scenario.}
\item{scenario. The index of the validation scenario.}
\item{dataset. The dataset index.}
}
}
\examples{

head(example_output)

visualize_parameter_group(
  example_output,
  pars = "lambda",
  theta_scenario = "1",
  scenarios = 1:2
)

}
