% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_val_sum.R
\docType{data}
\name{example_val_sum}
\alias{example_val_sum}
\title{example_val_sum: Example summaries of validated data}
\format{
A dataframe with 2 rows and 4 columns
}
\source{
Data was simulated using \link{simulate_validatedData} and summarized
using \link{summarize_n_validated}.
}
\description{
Example output from \link{summarize_n_validated}.
}
\details{
\itemize{
\item{theta_scenario. The ID of the classifier scenario.}
\item{scenario. The validation scenario index.}
\item{n_validated. The expected number of validated calls in an average dataset simulated under each scenario.}
}
}
\examples{
example_val_sum
}
