% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L.R, R/helpers.R
\name{L}
\alias{L}
\title{L: A labeller function borrowed from Ben XX on stack exchange.}
\usage{
L(labels, multi_line = TRUE)

L(labels, multi_line = TRUE)
}
\arguments{
\item{labels}{The labels to be parsed.}

\item{multi_line}{Whether the to print facet labels on}
}
\value{
facet labels
}
\description{
A function that allows symbols (like parameters with indices) to be parsed
in ggplot facet labels
}
\examples{
library(dplyr)
library(ggplot2)

tib <- tibble(
  pars = paste0("lambda[", rep(1:2, each = 5), "]"),
  vals = rnorm(10)
)

tib \%>\%
  ggplot(aes(x = vals)) +
  geom_histogram() +
  facet_wrap(~pars, labeller = L)

}
\keyword{internal}
