% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_not_confirmable}
\alias{make_not_confirmable}
\title{Further masking of true species labels to mimic
whether a recording is confirmable}
\usage{
make_not_confirmable(masked_df, confirmable_limits)
}
\value{
A copy of the original dataframe, with some selected observations masked
because they could not be confirmed.
}
\description{
Further masking of true species labels to mimic
whether a recording is confirmable
}
\keyword{internal}
