% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_by_spp.R
\name{mask_by_spp}
\alias{mask_by_spp}
\title{mask_by_spp}
\usage{
mask_by_spp(data, props_to_val)
}
\arguments{
\item{data}{A dataframe containing the columns \code{site}, \code{visit}, \code{true_spp}, \code{id_spp}, \code{count}}

\item{props_to_val}{a vector containing the proportion of recordings to validate for each species}
}
\value{
A dataframe object that is a copy of the input \code{data} masked according to the validation design supplied by \code{props_to_val}.
}
\description{
mask_by_spp
}
\examples{
library(dplyr)

dat <- sim_dat()$full_df

head(dat)

dat <- dat \%>\% tidyr::uncount(weights = count, .remove = FALSE)
val_dat <- mask_by_spp(dat, props_to_val = c(rep(.1, 4), rep(.4, 4)))

val_dat$final_df \%>\% group_by(id_spp) \%>\% summarize(prop_vald = sum(!is.na(true_spp))/n())
}
