% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_FE.R
\name{mask_FE}
\alias{mask_FE}
\title{mask_FE}
\usage{
mask_FE(df, effort_prop, seed = NULL)
}
\arguments{
\item{df}{A dataframe object}

\item{effort_prop}{The proportion of observations from the first visit to be validated.}

\item{seed}{An optional see to make masking/validation simulation reproducible}
}
\value{
A dataframe object with \code{effort_prop} of the observations validated from the first visit to each site.
}
\description{
mask_FE
}
\examples{
library(dplyr)

dat <- sim_dat()$full_df

head(dat)

dat <- dat \%>\% tidyr::uncount(weights = count, .remove = FALSE)
FE_validated_data <- mask_FE(dat, effort_prop = .2, seed = 17)

head(dat)
head(FE_validated_data)
}
