% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_FE_all_visits.R
\name{mask_FE_all_visits}
\alias{mask_FE_all_visits}
\title{Mask a proportion of all visits: A function for simulating a fixed effort validation design.}
\usage{
mask_FE_all_visits(df, effort_prop, seed = NULL)
}
\arguments{
\item{df}{A dataframe object in the format  of \code{full_dfs} output from
\link{simulate_validatedData}}

\item{effort_prop}{The proportion of recordings to be randomly sampled from
the first visit for validation}

\item{seed}{An optional random seed to make masking reproducible}
}
\value{
A dataframe object that is a copy of the input \code{df}, but with the
appropriate level of effort according to a fixed effort validation design.
}
\description{
Mask a proportion of all visits: A function for simulating a fixed effort validation design.
}
\examples{
library(dplyr)

cd_data <- sim_dat()$full_df \%>\% tidyr::uncount(weights = count, .remove = FALSE)
FE_data <- mask_FE_all_visits(cd_data, effort_prop = 0.1)

head(cd_data)
head(FE_data)

FE_data \%>\% 
  group_by(site, visit) \%>\% 
  summarize(
    prop_validated = sum(!is.na(true_spp))/n()
  )
  

}
