% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_sum.R
\name{mcmc_sum}
\alias{mcmc_sum}
\title{MCMC_sum}
\usage{
mcmc_sum(out, thin = 1, truth)
}
\arguments{
\item{out}{A list of outputs from each chain in the MCMC fit. Each list entry should be a matrix where each row corresponds to the iteration and each column corresponds to a parameter. This is the default output format for NIMBLE objects, as well as samples from other PPLs.}

\item{thin}{An integer value to thin draws by. Default value is 1.}

\item{truth}{a vector containing the true data-generating values for each model parameter}
}
\value{
a dataframe object with each row corresponding to parameters and columns corresponding to numeric summaries such as mean, sd, standard quantiles and MCMC diagnostics.
}
\description{
MCMC_sum
}
\examples{

library(rstan)
library(coda)
library(dplyr)

x <- runif(30, -1, 1)
beta <- -1
y <- rnorm(x*beta, sd = 1)

code <- nimbleCode({
  # priors
  beta ~ dnorm(0, 2)
  sigma ~ T(dnorm(0, 2), 0, Inf)

  # likelihood
  for(i in 1:n) {
    y[i] ~ dnorm(x[i]*beta, sd = sigma)
  }
})

fit <- nimbleMCMC(code,
  constants = list(n = length(y)),
  data = list(x = x, y = y),
  inits = list(beta = rnorm(1, 0, 10), sigma = abs(rnorm(0, 10))),
  nchains = 3
)

mcmc_sum(fit, truth = c(beta, 1))

}
