% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sims.R
\name{run_sims}
\alias{run_sims}
\title{run_sims: conduct simulations easily}
\usage{
run_sims(
  data_list,
  zeros_list,
  DGVs,
  theta_scenario_id,
  parallel = TRUE,
  niter = 2000,
  nburn = floor(niter/2),
  thin = 1,
  nchains = 3,
  save_fits = FALSE,
  save_individual_summaries_list = FALSE,
  directory = here::here()
)
}
\arguments{
\item{data_list}{nested list of masked dataframes (datasets nested within scenarios --
this is the format of \code{masked_dfs} output from \link{simulate_validatedData})}

\item{zeros_list}{list of dataframes containing the site/visit/true_spp/id_spp combinations
where no calls were observed.}

\item{DGVs}{A named list with entries psi, lambda and theta containing
the true values of the respective parameters.}

\item{theta_scenario_id}{The classifier ID as an integer or string}

\item{parallel}{Should models be fit in parallel? Default value is TRUE.}

\item{niter}{Number of iterations per MCMC chain.}

\item{nburn}{Number of warmup iterations.}

\item{thin}{Thinning interval for the MCMC chains.}

\item{nchains}{The number of chains.}

\item{save_fits}{Should individual model fits be saved? This could require large
amounts of disk space if you are fitting many large models to big datasets.
Default value is FALSE.}

\item{save_individual_summaries_list}{Should summaries for individual model fits
be saved? While this requires much less space than \code{save_fits}, we still
recommend keeping this at the default value of FALSE. Only use it if you
anticipate that simulations may be interrupted.}

\item{directory}{The directory to save objects. Defaults to the current working directory.}
}
\value{
a dataframe with the summaries (from \link{mcmc_sum}) for all scenarios and datasets.
A copy of this output is also saved to the current working directory.
}
\description{
run_sims: conduct simulations easily
}
\examples{
# :::::::::::: Simulate data ::::::::::::: #
psi <- c(0.3, 0.6)
lambda <- c(11, 2)

nspecies <- length(psi)
nsites <- 30
nvisits <- 5

test_theta1 <- matrix(c(0.9, 0.1, 0.15, 0.85), byrow = TRUE, nrow = 2)
val_scenarios <- list(spp1 = c(.75, .5), spp2 = .5)

fake_data <- simulate_validatedData(
  n_datasets = 5,
  design_type = "BySpecies",
  scenarios = val_scenarios,
  nsites = nsites,
  nvisits = nvisits,
  nspecies = nspecies,
  psi = psi,
  lambda = lambda,
  theta = test_theta1,
  save_datasets = FALSE,
  save_masked_datasets = FALSE,
  directory = paste0(here::here("Testing"))
)

# ::::::::::::: run simulations on sim'd data ::::::::::: #

\dontrun{ # turn off during check() to speed it up
out <- run_sims(
  data_list = fake_data$masked_dfs,
  zeros_list = fake_data$zeros,
  DGVs = list(lambda = lambda, psi = psi, theta = test_theta1),
  theta_scenario_id = 1,
  parallel = TRUE,
  niter = 1000,
  nburn = 500,
  thin = 1,
  save_fits = FALSE,
  save_individual_summaries_list = FALSE,
  directory = here::here()
)
}
}
