% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_validatedData.R
\name{simulate_validatedData}
\alias{simulate_validatedData}
\title{Simulate many datasets under candidate validation designs}
\usage{
simulate_validatedData(
  n_datasets,
  design_type = c("BySpecies", "FixedPercent"),
  scenarios,
  nsites,
  nspecies,
  nvisits,
  psi,
  lambda,
  theta,
  save_datasets = FALSE,
  save_masked_datasets = FALSE,
  directory = here::here()
)
}
\arguments{
\item{n_datasets}{The number of datasets you would like to have simulated. Each of these
simulated datasets will be subjected to all candidate validation designs.}

\item{design_type}{Character string, either "BySpecies" for a stratified-by-species design,
or "FixedPercentage" for a fixed effort design (see Oram et al., in review for more details on each of these)}

\item{scenarios}{if \code{design_type = "BySpecies"}, the \code{scenarios} argument must be a list with each entry corresponding
to the potential levels of effort for a particular autoID label. If
\code{design_type == "FixedPercent"}, then the \code{scenarios} argument must be a vector with each entry corresponding to
a potential percent of calls to be sampled from the first visit at each site. See vignette for an example.}

\item{nsites}{number of sites in each dataset}

\item{nspecies}{size of the species assemblage}

\item{nvisits}{the number of visits to each site. Note that these simulations assume a balanced design.}

\item{psi}{a vector of length nspecies with the assumed occurrence probabilities for each species}

\item{lambda}{a vector of length nspecies with the assumed relative activity levels for each species.
Make sure the order is correct and matches psi.}

\item{theta}{a matrix containing the (mis)classification probabilities. The rows of this matrix must sum
to 1. See vignette for an example.}

\item{save_datasets}{logical. If TRUE, the datasets without any masking of true species labels (i.e.,
corresponding to complete validation of all recordings) will be saved. Default value is FALSE.}

\item{save_masked_datasets}{logical. If TRUE, the masked datasets (i.e., the simulated datasets with
partial validation according to the simulation scenario) will be saved. This means that there will be
n_datasets x nrow(scenarios_dataframe) datasets saved: one for each dataset under each validation scenario.
Default value is FALSE.}

\item{directory}{character. Required if save_datasets = TRUE or save_masked_datasets = TRUE. This is where the
datasets will be saved. By default, the current working directory (i.e., here::here()) will be used.}
}
\value{
A list containing three elements:
\enumerate{
\item \code{full_datasets}: A list of length n_datasets with unmasked datasets (i.e., full validation of all recordings).
If \code{save_datasets = TRUE}, then these will be saved individually in \code{directory} as dataset_n.rds, where n
is the dataset number.
\item \code{zeros}: A list of length n_datasets containing all of the site-visits where no recordings of a certain
classification were observed. For example, if, in dataset 10, there were no calls from species 1 that were
classified as  3 on visit 4 to site 156, then the 10th entry of this list would contain a dataset with
a row corresponding tosite = 156, visit = 4, true_spp = 1, id_spp = 3, with count = 0. These zeros are
necessary for housekeeping in the model-fitting process. If \code{save_datasets = TRUE}, the zeros for each
each dataset will be saved in \code{directory} individually as zeros_in_dataset_n.rds, where
n is the dataset number.
\item \code{masked_dfs}: A nested list containing each dataset masked under each scenario. masked_dfs\[\[9\]\]\[\[27\]\] contains
dataset 27, assuming validation scenario 9. If \code{save_masked_datasets = TRUE}, then each dataset/scenario
scenario combination is saved individually in \code{directory} as dataset_n_masked_under_scenario_s.rds,
where n is the dataset number and s is the scenario number.
}
}
\description{
Simulate many datasets under candidate validation designs
}
\examples{

psi <- c(0.3, 0.6)
lambda <- c(11, 2)

nspecies <- length(psi)
nsites <- 30
nvisits <- 5

test_theta1 <- matrix(c(0.9, 0.1, 0.15, 0.85), byrow = TRUE, nrow = 2)
val_scenarios <- list(spp1 = c(.75, .5), spp2 = .5)

fake_data <- simulate_validatedData(
  n_datasets = 5,
  design_type = "BySpecies",
  scenarios = val_scenarios,
  nsites = nsites,
  nvisits = nvisits,
  nspecies = nspecies,
  psi = psi,
  lambda = lambda,
  theta = test_theta1,
  save_datasets = FALSE,
  save_masked_datasets = FALSE,
  directory = paste0(here::here("Testing"))
)


}
