% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_parameters.R
\name{visualize_parameter_group}
\alias{visualize_parameter_group}
\title{visualize_parameter_group}
\usage{
visualize_parameter_group(
  sim_summary,
  pars,
  theta_scenario,
  scenarios,
  convergence_threshold = 1.1
)
}
\arguments{
\item{sim_summary}{Summary output in the format of \code{run_sims()}.}

\item{pars}{The parameters to be visualized.}

\item{theta_scenario}{The theta scenario IDs. This should match the \code{theta_scenario_id} argument of\code{run_sims()}.}

\item{scenarios}{Scenarios to be visualized.}

\item{convergence_threshold}{If the Gelman-Rubin statistic is below this value,
consider an MCMC to have converged. Default value is 1.1, but we recommend 1.05.}
}
\value{
A ggplot object summarizing simulation results.
}
\description{
Visualize simulation results for entire groups of parameters
}
\examples{

visualize_parameter_group(
  example_output,
  pars = "lambda",
  theta_scenario = "1",
  scenarios = 1:2
)

}
