% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_parameters.R
\name{visualize_single_parameter}
\alias{visualize_single_parameter}
\title{Visualize your simulation study results}
\usage{
visualize_single_parameter(
  sim_summary,
  par,
  theta_scenario,
  scenarios,
  convergence_threshold = 1.1
)
}
\arguments{
\item{sim_summary}{Summary output in the format of \code{run_sims()}.}

\item{par}{The parameter(s) to be visualized.}

\item{theta_scenario}{The theta scenario IDs.}

\item{scenarios}{Scenarios to be visualized.}

\item{convergence_threshold}{If the Gelman-Rubin statistic is below this value,
consider an MCMC to have converged. Default value is 1.1, but we recommend 1.05.}
}
\value{
A ggplot object summarizing simulation results.
}
\description{
Visualize your simulation study results
}
\examples{


visualize_single_parameter(
  example_output,
  par = "lambda[1]",
  theta_scenario = "1",
  scenarios = 1:2,
  convergence_threshold = 1.05
)

visualize_parameter_group(
  example_output,
  pars = "lambda"
  theta_scenario = "1",
  scenarios = 1:2
)

}
