% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{L}
\alias{L}
\title{L: A labeller function from Ben Bolker on Stack Overflow}
\usage{
L(labels, multi_line = TRUE)
}
\description{
A function that allows symbols (like parameters with indices) to be parsed
in ggplot facet labels
}
\keyword{internal}
