% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_FE.R
\name{mask_FE}
\alias{mask_FE}
\title{mask_FE: A function for simulating a fixed effort validation design.}
\usage{
mask_FE(df, effort_prop, seed = NULL)
}
\arguments{
\item{df}{A dataframe object in the format  of \code{full_dfs} output from
\link{simulate_validatedData}}

\item{effort_prop}{The proportion of recordings to be randomly sampled from
the first visit for validation}

\item{seed}{An optional random seed to make masking reproducible}
}
\value{
A dataframe object that is a copy of the input \code{df}, but with the
appropriate level of effort according to a fixed effort validation design.
}
\description{
mask_FE: A function for simulating a fixed effort validation design.
}
\examples{
cd_data <- sim_dat()$full_df \%>\% tidyr::uncount(weights = count, .remove = FALSE)

FE_data <- mask_FE(cd_data, effort_prop = 0.2)

head(cd_data)
head(FE_data)

}
