% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_FE.R, R/mask_FE_h.R
\name{mask_FE}
\alias{mask_FE}
\title{mask_FE: A function for simulating a fixed effort validation design.}
\usage{
mask_FE(df, effort_prop, seed = NULL)

mask_FE(df, effort_prop, seed = NULL)
}
\arguments{
\item{df}{A dataframe object in the format  of \code{full_dfs} output from
\link{simulate_validatedData}}

\item{effort_prop}{The proportion of recordings to be randomly sampled from
the each site visit for validation.}

\item{seed}{An optional random seed to make masking reproducible}
}
\value{
A dataframe object that is a copy of the input \code{df}, but with the
appropriate level of effort according to a fixed effort validation design.

A dataframe object that is a copy of the input \code{df}, but with the
appropriate level of effort according to a fixed effort validation design.
}
\description{
This function implements a fixed effort design to be used when heterogenous
relative activity is expected. Rather than validating a random x\% from the
first visit, which assumes that the first visit is representative of all
others in terms of the number of recordings that are expected, this design
randomly selects y\% from each visit. This is the design used in the Oregon-
Washington dataset that was previously analyzed by Oram et al. (in
submission).
}
\examples{
cd_data <- sim_dat()$full_df \%>\% tidyr::uncount(weights = count, .remove = FALSE)

FE_data <- mask_FE(cd_data, effort_prop = 0.2)

head(cd_data)
head(FE_data)

cd_data <- sim_dat()$full_df \%>\% tidyr::uncount(weights = count, .remove = FALSE)

FE_data <- mask_FE(cd_data, effort_prop = 0.2)

head(cd_data)
head(FE_data)

}
