% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bias_vs_calls.R
\name{plot_bias_vs_calls}
\alias{plot_bias_vs_calls}
\title{plot_bias_vs_calls.R}
\usage{
plot_bias_vs_calls(
  sim_summary,
  calls_summary,
  pars = NULL,
  regex_pars = NULL,
  theta_scenario,
  scenarios,
  convergence_threshold = 1.1
)
}
\arguments{
\item{sim_summary}{Simulation summaries in the format output from \code{run_sims}. This means a dataframe object with columns for \code{parameter}, (posterior) \code{Mean}, \code{truth}, \code{theta_scenario}, (validation) \code{scenario}, and \code{dataset} are all available.}

\item{calls_summary}{A summary of the validated data as output from \code{summarize_n_validated}. Must contain the columns \code{theta_scenario} and \code{scenario}}

\item{pars}{A character vector containing the parameters to be visualized. Use one of the following: "psi", "lambda", or "theta".}

\item{regex_pars}{A string containing the name of a parameter group (e.g., "lambda") to be visualized.}

\item{theta_scenario}{The classifier ID to be visualized.}

\item{scenarios}{The ID's corresponding to the validation scenarios to be visualized. Typically a numeric vector (e.g. 1:5)}

\item{convergence_threshold}{The threshold  value for the Gelman-Rubin statistic. Below this value, a model is considered 'converged' for the purpose of including it in visualizations.}
}
\value{
A ggplot object  of average estimation error vs the number of validated calls under each scenario.
}
\description{
plot_bias_vs_calls.R
}
\examples{

s1 <- readRDS("Testing/FixedEffortExample/ThetaFE/summary_df_for_scenario_1.rds")
s2 <- readRDS("Testing/FixedEffortExample/ThetaFE/summary_df_for_scenario_2.rds")

s <- bind_rows(s1, s2)
plot_bias_vs_calls(s, calls_summary)
}
